quest half done
updating ui is complete, the home button seems fine. did not implement a button to go into this scene. as its gonna be an overlay afterwards.
press z or x or c to update each section of mission. press r for random generate mission below. the login is usually the first mission so do not need to randomly generate (following other mobile games as reference)

needed to do:
-update exp and coin with database
  when press claim
-go shortcut need to do camera movement or something to go to designated location,
-update the button to a complete image when completed mission
-trigger event when mission meet
-make this an overlay