saving my stuff
last time, clicked unity cloud.... and got rid of my safe.... whats worse was on github there was merge conflcit which i couldnt even reverse it.... i finished it manually, by pulling the quest 70 done part and redo the register and stuff.... then got it working. its just that the save is too old... so what i do is reclonem and copy and paste this locally saved file in here. hope it works