figured out overlay code.
cant test it..... the server seems down or something. have to wait for product owner to put back server so i can actually test it. and also take screenshots of tutorial and update it.
in case if he changed his mind, as well, so i didn't overwrite the previous code, just commented it below in case if i needed it.